requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
psycopg==3.1.13
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
